DPS-1616: Reset this library back to version 5.0.2
This library is hosted on both github and gitlab with diverging history.
It's only used by evw-integration-service (as far as I can tell) which
uses version 5.0.2. Therefore I'm setting master back to this version
before moving it to ACP and this is the only solution that seems to make
sense. I've updated the README.md on github to say "use the gitlab version"